[FEATURE-5] Move files to FSD folders
