+++ src/Host/ClientApp/src/entities/Link.tsx
