+++ src/Host/ClientApp/src/features/Authorization.ts
@@ -1,4 +1,4 @@
-import { localhostApi } from "../services/localhostApi";
+import { localhostApi } from "../shared/localhostApi";
 
 export interface AuthorizationState {
     data: AuthorizationData,
