+++ src/Host/ClientApp/src/shared/AppTable.tsx
